Extract tooltip-related logic into a hook
